backend/producer/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2      
Jinja2==3.1.2
kafka==1.3.5
kafka-python==2.0.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
python-dotenv==0.19.1
Werkzeug==2.2.3
zipp==3.14.0
